refactor: factory 패키지 삭제
- 팩토리 패턴을 클래스 내에서 구현합니다.
- 도메인의 경우 테스트를 위해 생성자를 public으로 설정합니다.
- 이외의 클래스의 생성자는 private로 설정한 뒤 static 메서드를 통해 생성하도록 수정합니다.
