+++ src/main/java/lotto/controller/WinningController.java
@@ -1,6 +1,5 @@
 package lotto.controller;
 
-import lotto.factory.WinningGeneratorFactory;
 import lotto.service.generator.WinningGenerator;
 import lotto.view.InputView;
 import lotto.view.OutputView;
@@ -23,7 +22,7 @@ public class WinningController {
         outputView.printlnMessage(PrintMessage.INPUT_LOTTO_WINNING_NUMBER);
         String lottoWinning = inputView.inputUser();
 
-        return WinningGeneratorFactory.create(lottoWinning);
+        return WinningGenerator.create(lottoWinning);
     }
 
     public WinningGenerator getWinningGenerator() {
