+++ src/main/java/lotto/domain/Wallet.java
@@ -12,6 +12,10 @@ public class Wallet {
         this.money = money;
     }
 
+    public static Wallet create(Integer money) {
+        return new Wallet(money);
+    }
+
     private void validate(Integer money) {
         if (money % MIN_MONEY_UNIT != MIN_MONEY_UNIT_CHECK_NUMBER) {
             throw new IllegalArgumentException();
