--- src/main/java/lotto/factory/BonusFactory.java
@@ -1,10 +1,0 @@
-package lotto.factory;
-
-import lotto.domain.Bonus;
-
-public class BonusFactory {
-
-    public static Bonus create(Integer bonus) {
-        return new Bonus(bonus);
-    }
-}
