--- src/main/java/lotto/factory/LottoFactory.java
@@ -1,11 +1,0 @@
-package lotto.factory;
-
-import java.util.List;
-import lotto.domain.Lotto;
-
-public class LottoFactory {
-
-    public static Lotto create(List<Integer> number) {
-        return new Lotto(number);
-    }
-}
