--- src/main/java/lotto/factory/LottoGeneratorFactory.java
@@ -1,10 +1,0 @@
-package lotto.factory;
-
-import lotto.service.generator.LottoGenerator;
-
-public class LottoGeneratorFactory {
-
-    public static LottoGenerator create(Integer lottoCount) {
-        return new LottoGenerator(lottoCount);
-    }
-}
