--- src/main/java/lotto/factory/PaymentFactory.java
@@ -1,10 +1,0 @@
-package lotto.factory;
-
-import lotto.service.Payment;
-
-public class PaymentFactory {
-
-    public static Payment create(String money) {
-        return new Payment(money);
-    }
-}
