--- src/main/java/lotto/factory/WalletFactory.java
@@ -1,10 +1,0 @@
-package lotto.factory;
-
-import lotto.domain.Wallet;
-
-public class WalletFactory {
-
-    public static Wallet create(Integer money) {
-        return new Wallet(money);
-    }
-}
