--- src/main/java/lotto/factory/WinningGeneratorFactory.java
@@ -1,10 +1,0 @@
-package lotto.factory;
-
-import lotto.service.generator.WinningGenerator;
-
-public class WinningGeneratorFactory {
-
-    public static WinningGenerator create(String winning) {
-        return new WinningGenerator(winning);
-    }
-}
