+++ src/main/java/lotto/service/generator/LottoGenerator.java
@@ -3,7 +3,6 @@ package lotto.service.generator;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
-import lotto.factory.LottoFactory;
 import lotto.domain.Lotto;
 
 public class LottoGenerator {
@@ -15,8 +14,12 @@ public class LottoGenerator {
     private final Integer lottoCount;
     private final List<Lotto> lottoTicket = new ArrayList<>();
 
-    public LottoGenerator(Integer lottoCount) {
+    private LottoGenerator(Integer lottoCount) {
         this.lottoCount = lottoCount;
+    }
+
+    public static LottoGenerator create(Integer lottoCount) {
+        return new LottoGenerator(lottoCount);
     }
 
     private List<Integer> generateNumber() {
@@ -24,7 +27,7 @@ public class LottoGenerator {
     }
 
     private Lotto changeNumberToLotto(List<Integer> number) {
-        return LottoFactory.create(number);
+        return Lotto.create(number);
     }
 
     public List<Lotto> getLottoTicket() {
